Fix integer out of range error on Postgres 11
On PostgreSQL 11 (haven't tested on other versions) an integer out of range error is thrown due to casting to INT:
```
2018-12-03 03:22:25: python.d ERROR: postgres: socket: update() unhandled exception: integer out of range
```
Casting to BIGINT solves this error.
The PostgreSQL Docs do not list wal_block_size as bigint, so I do not fully understand why this is required.
https://www.postgresql.org/docs/11/functions-info.html

+++ collectors/python.d.plugin/postgres/postgres.chart.py
@@ -261,7 +261,7 @@ WHERE application_name IS NOT NULL;
     'REPSLOT_FILES': """
 WITH wal_size AS (
   SELECT
-    current_setting('wal_block_size')::INT * setting::INT AS val
+    current_setting('wal_block_size')::BIGINT * setting::BIGINT AS val
   FROM pg_settings
   WHERE name = 'wal_segment_size'
   )
